backend/api/requirements.txt
pandas==1.5.3
pyodbc==5.0.1
python-dotenv==0.21.1
uvicorn==0.30.1
fastapi==0.111.0
numpy==1.26.4
